Configurable stop signal for docker images
Snowplow apps try to do graceful shutdown and exit when they are ready.
However some 3rd-pary libraries try to do their own graceful shutdown
which interferes with Snowplow's graceful shutdown.

Our solution is for the Snowplow app to start graceful shutdown on a
non-standard signal.  The signal we choose must cause the snowplow app
to start shutting down, but it must be ignored by 3rd party libraries.

This commit makes the STOPSIGNAL configurable so we can configure Docker
to stop the app with our chosen signal.