Java 21 for snowplow docker images (#10)
For the last few years, Snowplow has used Java 11.  For most apps we
provide an image based off unbuntu, plus a distroless image, based off
the google distroless project, which in turn was based off debian.

This PR upgrades to Java 21, the most recent LTS version. Our new app
images use the following 3rd party base images:
- `eclipse-temurin:21-jre-noble`.  This base image uses ubuntu 24.04
  (noble) and the eclipse temurin build of the JRE.
- `gcr.io/distroless/java21-debian12`. This distroless base image is
  extended from debian 12, and also uses the temurin build of the JRE

Before this PR, we had a much more complicated way of customizing the
distroless image. But this was done mainly to exclude libssl (plus some
other libs). The new distroless base image already exludes libssl, so it
does not feel worth the effort of maintaining a complicated
customization. So I have reverted back to using a simpler base image
definition for the distroless image. The downside is we don't exclude
quite so many libs as we did before, e.g. libexpat, which we previously
excluded when there was an old CVE on an earlier version.